refactor to make project structure more homogenous
