+++ src/main/java/edu/aau/groupc/canteenbackend/DBEntity.java
@@ -1,0 +1,4 @@
+package edu.aau.groupc.canteenbackend;
+
+public interface DBEntity {
+}
