+++ src/main/java/edu/aau/groupc/canteenbackend/auth/Auth.java
@@ -1,6 +1,6 @@
 package edu.aau.groupc.canteenbackend.auth;
 
-import edu.aau.groupc.canteenbackend.entities.DBEntity;
+import edu.aau.groupc.canteenbackend.DBEntity;
 import lombok.Data;
 
 import javax.persistence.*;
