+++ src/main/java/edu/aau/groupc/canteenbackend/config/ResponseCorsFilter.java
@@ -1,5 +1,6 @@
 package edu.aau.groupc.canteenbackend.config;
 
+import org.springframework.lang.NonNull;
 import org.springframework.stereotype.Component;
 import org.springframework.web.filter.OncePerRequestFilter;
 
@@ -10,10 +11,10 @@ import javax.servlet.http.HttpServletResponse;
 import java.io.IOException;
 
 @Component
-public class CanteenCorsFilter extends OncePerRequestFilter {
+public class ResponseCorsFilter extends OncePerRequestFilter {
 
     @Override
-    protected void doFilterInternal(HttpServletRequest httpServletRequest,
+    protected void doFilterInternal(@NonNull HttpServletRequest httpServletRequest,
                                     HttpServletResponse httpServletResponse,
                                     FilterChain filterChain)
             throws ServletException, IOException {
