+++ src/main/java/edu/aau/groupc/canteenbackend/dish/services/DishService.java
@@ -1,7 +1,7 @@
-package edu.aau.groupc.canteenbackend.services;
+package edu.aau.groupc.canteenbackend.dish.services;
 
-import edu.aau.groupc.canteenbackend.dao.DishRepository;
-import edu.aau.groupc.canteenbackend.entities.Dish;
+import edu.aau.groupc.canteenbackend.dish.repositories.DishRepository;
+import edu.aau.groupc.canteenbackend.dish.Dish;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.ResponseEntity;
 import org.springframework.http.HttpStatus;
