--- src/main/java/edu/aau/groupc/canteenbackend/entities/DBEntity.java
@@ -1,4 +1,0 @@
-package edu.aau.groupc.canteenbackend.entities;
-
-public interface DBEntity {
-}
