+++ src/test/java/edu/aau/groupc/canteenbackend/order/services/OrderHasDishServiceTest.java
@@ -1,11 +1,11 @@
 package edu.aau.groupc.canteenbackend.order.services;
 
-import edu.aau.groupc.canteenbackend.entities.Dish;
+import edu.aau.groupc.canteenbackend.dish.Dish;
 import edu.aau.groupc.canteenbackend.orders.Order;
 import edu.aau.groupc.canteenbackend.orders.OrderHasDish;
 import edu.aau.groupc.canteenbackend.orders.services.IOrderHasDishService;
 import edu.aau.groupc.canteenbackend.orders.services.IOrderService;
-import edu.aau.groupc.canteenbackend.services.IDishService;
+import edu.aau.groupc.canteenbackend.dish.services.IDishService;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
